+++ src/compiler/emitter/mod.rs
@@ -12,5 +12,7 @@ mod lines_3000_3500;
 pub use lines_3000_3500::*;
 mod lines_3500_4000;
 pub use lines_3500_4000::*;
-mod lines_4000_end;
-pub use lines_4000_end::*;
+mod lines_4000_4500;
+pub use lines_4000_4500::*;
+mod lines_4500_end;
+pub use lines_4500_end::*;
